+++ cx-portal/src/components/pages/Admin/components/RegistrationRequests/components/RequestList/index.tsx
@@ -27,6 +27,7 @@ import { updateApplicationRequestSelector } from 'features/control/updatesSlice'
 import { ApplicationRequest } from 'features/admin/applicationRequestApiSlice'
 import { RegistrationRequestsTableColumns } from '../../registrationTableColumns'
 import { GridCellParams } from '@mui/x-data-grid'
+import './RequestListStyle.scss'
 
 export const RequestList = ({
   fetchHook,
@@ -57,6 +58,31 @@ export const RequestList = ({
   const dispatch = useDispatch()
   const [refresh, setRefresh] = useState<number>(0)
   const searchInputData = useSelector(updateApplicationRequestSelector)
+  const [group, setGroup] = useState<string>('')
+
+  const setView = (e: React.MouseEvent<HTMLInputElement>) => {
+    const viewValue = e.currentTarget.value
+    setGroup(viewValue)
+    setRefresh(Date.now())
+  }
+
+  const filterView = [
+    {
+      buttonText: t('content.admin.registration-requests.filter.all'),
+      buttonValue: 'inactive',
+      onButtonClick: setView,
+    },
+    {
+      buttonText: t('content.admin.registration-requests.filter.review'),
+      buttonValue: '',
+      onButtonClick: setView,
+    },
+    {
+      buttonText: t('content.admin.registration-requests.filter.closed'),
+      buttonValue: 'active',
+      onButtonClick: setView,
+    },
+  ]
 
   const columns = RegistrationRequestsTableColumns(
     useTranslation,
@@ -78,12 +104,12 @@ export const RequestList = ({
   }
 
   return (
-    <section id="identity-management-id">
+    <section id="registration-section-id">
       <PageLoadingTable<ApplicationRequest>
         searchExpr={searchExpr}
         rowHeight={80}
         onCellClick={onTableCellClick}
-        toolbarVariant={'ultimate'}
+        toolbarVariant={'searchAndFilter'}
         hasBorder={false}
         columnHeadersBackgroundColor={'transparent'}
         searchPlaceholder={t('global.table.searchName')}
@@ -101,6 +127,11 @@ export const RequestList = ({
         fetchHookRefresh={refresh}
         getRowId={(row: { [key: string]: string }) => row.applicationId}
         columns={columns}
+        descriptionText={`${t(
+          'content.admin.registration-requests.introText1'
+        )}${t('content.admin.registration-requests.introText2')}`}
+        defaultFilter={group}
+        filterViews={filterView}
       />
     </section>
   )